% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_systems.R
\name{classification_systems}
\alias{classification_systems}
\title{...}
\usage{
classification_systems(q, system_id = NULL, params_list = list())
}
\arguments{
\item{q}{...}

\item{system_id}{...}

\item{params_list}{...}
}
\value{
...
}
\description{
...
}
