% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{doc_query}
\alias{doc_query}
\title{Document utils functions}
\usage{
doc_query(d)
}
\arguments{
\item{d}{\code{RLCCSDocument} object}
}
\value{
a \code{RLCCSQuery} object with the predecessor subclass with the
 fields used in the request.
}
\description{
Document utils functions
}
