% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{extensions}
\alias{extensions}
\alias{endpoint}
\alias{before_request}
\alias{after_response}
\alias{content_response}
\alias{check_query_verb}
\alias{check_subclass}
\alias{subclass}
\alias{omit_query_params}
\title{Extension development functions}
\usage{
endpoint(q)

before_request(q)

after_response(q, res)

content_response(res, status_codes, content_types)

check_query_verb(q, verbs, msg = NULL)

check_subclass(x, subclasses)

subclass(x)

omit_query_params(q, names)
}
\arguments{
\item{q}{a \code{RLCCSQuery} object.}

\item{res}{a \code{httr} \code{response} object.}

\item{status_codes}{a \code{character} vector with successful
status codes.}

\item{content_types}{a \code{character} vector with all acceptable
responses' content type.}

\item{verbs}{a \code{character} vector with allowed HTTP request methods}

\item{msg}{a \code{character} with a personalized error message}

\item{x}{either a \code{RLCCSQuery} object expressing a LCCS query
criteria or any \code{RLCCSDocument}.}

\item{subclasses}{a \code{character} vector with all allowed S3 subclasses}

\item{names}{a \code{character} vector with the names do omit.}
}
\value{
A \code{character} endpoint value for \code{endpoint()} function.
A \code{RLCCSQuery} object for \code{before_request()} and
\code{after_response()} functions.

The \code{content_response()} function returns a \code{list} data structure
representing the JSON file received in HTTP response
}
\description{
...
}
\section{Functions}{
\itemize{
\item \code{content_response}: The \code{content_response} function checks if the request's
response is in accordance with the allowed status codes and content-types.
It returns the parsed content response.

\item \code{check_query_verb}: The \code{check_query_verb()} function allows you to define which HTTP
verbs are allowed. It is useful for establishing which verbs will be
supported by an extension.

\item \code{check_subclass}: The \code{check_subclass()} function specifies which type of query
objects (\code{RLCCSQuery}) or document objects (\code{RLCCSDocument})
are expected in the function extension.

\item \code{subclass}: The \code{subclass()} function returns a \code{character} representing the
subclass name of either \code{RLCCSQuery} or \code{RSTACDocument} S3 classes.

\item \code{omit_query_params}: The \code{omit_query_params()} function was created to omit the paths that
are defined as query parameters to simplify the creation of a query.
Therefore, use this method only in endpoints that specify a parameter in
their paths.
}}

