% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccs.R
\name{lccs}
\alias{lccs}
\title{Endpoint functions}
\usage{
lccs(url = NULL, .token = NULL)
}
\arguments{
\item{url}{a \code{character} informing the base url of a
LCCS-WS (LCCS-WS-SPEC 0.6)}

\item{.token}{a \code{character} informing the authentication token in the
BDC-OAuth service. This token is used for the service's administrative
operations through the HTTP verbs POST, PUT, DELETE.}
}
\value{
A \code{RLCCSQuery} object with the subclass \code{lccs} containing all
request parameters to be provided to API service.
}
\description{
The \code{lccs} function implements \code{/} API
endpoint. It prepares search fields parameters to be provided to
a LCCS-WS. This endpoint should return a \code{RLCCSDocument} containing links
to LCCS-WS API root and Classification Systems published.
}
\examples{
\donttest{
lccs("https://brazildatacube.dpi.inpe.br/dev/lccs/") \%>\%
  get_request()
}
}
