% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{mappings}
\alias{mappings}
\title{...}
\usage{
mappings(q, system_id_source, system_id_target = NULL, params_list = list())
}
\arguments{
\item{q}{...}

\item{system_id_source}{...}

\item{system_id_target}{...}

\item{params_list}{...}
}
\value{
...
}
\description{
...
}
