% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_request}
\alias{get_request}
\alias{post_request}
\alias{put_request}
\alias{delete_request}
\title{LCCS API request functions}
\usage{
get_request(q, ...)

post_request(q, ..., encode = c("json", "multipart", "form"))

put_request(q, ..., encode = c("json", "multipart", "form"))

delete_request(q, ...)
}
\arguments{
\item{q}{a \code{RLCCSQuery} object expressing a LCCS query
criteria.}

\item{...}{...}

\item{encode}{a \code{character} informing the request body
Content-Type. Accepted types are \code{'json'} (\code{'application/json'}),
\code{'form'} (\code{'application/x-www-form-urlencoded'}),
and \code{'multipart'} (\code{'multipart/form-data'}). Defaults to
\code{'json'}.}
}
\value{
...
}
\description{
The \code{get_request} ...

The \code{post_request} ...

The \code{put_request} ...

The \code{delete_request} ...
}
