% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_formats.R
\name{style_formats}
\alias{style_formats}
\title{...}
\usage{
style_formats(q, style_format_id = NULL, params_list = list())
}
\arguments{
\item{q}{...}

\item{style_format_id}{...}

\item{params_list}{...}
}
\value{
...
}
\description{
...
}
