% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{styles}
\alias{styles}
\title{...}
\usage{
styles(q, style_id = NULL, style_file = NULL, params_list = list())
}
\arguments{
\item{q}{...}

\item{params_list}{...}

\item{system_id}{...}
}
\value{
...
}
\description{
...
}
