% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{classes}
\alias{classes}
\title{...}
\usage{
classes(q, class_id = NULL, params_list = list())
}
\arguments{
\item{q}{...}

\item{params_list}{...}

\item{system_id}{...}
}
\value{
...
}
\description{
...
}
