% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_request}
\alias{get_request}
\alias{post_request}
\alias{put_request}
\alias{delete_request}
\title{LCCS API request functions}
\usage{
get_request(q, ...)

post_request(q, ..., encode = c("json", "multipart", "form"))

put_request(q, ..., encode = c("json", "multipart", "form"))

delete_request(q, ...)
}
\arguments{
\item{q}{a \code{RLCCSQuery} object expressing a LCCS query
criteria.}

\item{...}{extra parameters to be used in \code{httr} HTTP verbs functions
(GET, POST, PUT, DELETE)}

\item{encode}{a \code{character} informing the request body
Content-Type. Accepted types are \code{'json'} (\code{'application/json'}),
\code{'form'} (\code{'application/x-www-form-urlencoded'}),
and \code{'multipart'} (\code{'multipart/form-data'}). Defaults to
\code{'json'}.}
}
\value{
Returns an \code{RLCCSDocument} The contents of the class may vary depending
on the operation being
}
\description{
The \code{get_request} is used to express HTTP-GET verb in LCCS-WS
endpoints

The \code{post_request} is used to express HTTP-POST verb in LCCS-WS
endpoints. This require a BDC-OAuth token

The \code{put_request} is used to express HTTP-PUT verb in LCCS-WS
endpoints. This require a BDC-OAuth token

The \code{delete_request} is used to express HTTP-DELETE verb in LCCS-WS
endpoints. This require a BDC-OAuth token
}
