% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{RLCCSQuery}
\alias{RLCCSQuery}
\title{...}
\usage{
RLCCSQuery(base_url, params = list(), subclass)
}
\arguments{
\item{base_url}{...}

\item{params}{...}

\item{subclass}{...}
}
\value{
a class ...
}
\description{
...
}
