% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{RLCCSQuery}
\alias{RLCCSQuery}
\title{...}
\usage{
RLCCSQuery(
  base_url,
  params = list(),
  subclass,
  query_type = "key",
  encode = NULL
)
}
\arguments{
\item{base_url}{...}

\item{params}{...}

\item{subclass}{...}

\item{query_type}{...}

\item{encode}{...}
}
\value{
a class ...
}
\description{
...
}
